WIP: use seeds for initial hash table in pmh
in the papers that prove that pmh works in linear time, the initial
hash also should have a random seed.

as an experiment, in this commit, we make the first hash also seeded,
and try several seeds. we reject seeds that have a high score for
sum of bucket size squared, and don't do more than a thousand seeds

currently this seems to fail tests unfortuantely, it seems to slow
down the lookup enough that it doesn't beat std::map on small sets
also I think we are failing some unit tests...