fix container typedefs in frozen::unordered_map
In C++ std::unordered_map<U, V>, there are standard typedefs:

::key_type    U
::mapped_type V
::value_type  pair<U, V>

In frozen::unordered_map, value_type was V, and mapped_type was
omitted.

Usually, value_type is the result of *'ing an iterator.

In frozen::map, we did it the conventional way, there was only
a discrepancy in frozen::unordered_map. So I think it was just
an oversight and in this commit we fix it.

We add some tests also in the form of static_assert